requirements.txt
paramiko==2.8.0
python-dotenv